+++ rollup.config.js
@@ -6,6 +6,8 @@ import svelte from 'rollup-plugin-svelte';
 import babel from 'rollup-plugin-babel';
 import { terser } from 'rollup-plugin-terser';
 
+const production = !process.env.ROLLUP_WATCH;
+
 const output = {
     name: 'chart',
     dir: path.resolve(__dirname, 'dist'),
@@ -34,12 +36,13 @@ module.exports = [
             svelte({ hydratable: true }),
             resolve(),
             commonjs(),
-            babel({
-                ...babelConfig,
-                presets: [['@babel/env', { targets: '> 1%', corejs: 3, useBuiltIns: 'entry' }]],
-                plugins: ['babel-plugin-transform-async-to-promises']
-            }),
-            terser()
+            production &&
+                babel({
+                    ...babelConfig,
+                    presets: [['@babel/env', { targets: '> 1%', corejs: 3, useBuiltIns: 'entry' }]],
+                    plugins: ['babel-plugin-transform-async-to-promises']
+                }),
+            production && terser()
         ],
         external: ['underscore'],
         onwarn,
@@ -51,7 +54,7 @@ module.exports = [
     },
     {
         /* Server side rendered Svelte Chart Component */
-        input: path.resolve(__dirname, 'lib/Chart.svelte'),
+        input: path.resolve(__dirname, 'lib/Visualization.svelte'),
         plugins: [
             svelte({ generate: 'ssr', hydratable: true }),
             resolve(),
@@ -65,7 +68,7 @@ module.exports = [
         onwarn,
         output: {
             format: 'umd',
-            entryFileNames: 'Chart_SSR.js',
+            entryFileNames: 'Visualization_SSR.js',
             ...output
         }
     },
@@ -79,7 +82,7 @@ module.exports = [
                 presets: [['@babel/env', { targets: '> 1%', corejs: 3, useBuiltIns: 'entry' }]],
                 plugins: ['babel-plugin-transform-async-to-promises']
             }),
-            terser()
+            production && terser()
         ],
         external: ['underscore'],
         onwarn,
@@ -99,7 +102,7 @@ module.exports = [
                 presets: [['@babel/env', { targets: '> 1%', corejs: 3, useBuiltIns: 'entry' }]],
                 plugins: ['babel-plugin-transform-async-to-promises']
             }),
-            terser()
+            production && terser()
         ],
         output: {
             name: 'embed',
